Rename RightAlignedButtonRow to SectionButtonRow
