--- context/app/static/js/shared-styles/sections/RightAlignedButtonRow/index.js
@@ -1,4 +1,0 @@
-import RightAlignedButtonRow, { StyledButtonRow, BottomAlignedTypography } from './RightAlignedButtonRow';
-
-export { StyledButtonRow, BottomAlignedTypography };
-export default RightAlignedButtonRow;
