+++ context/app/static/js/shared-styles/sections/SectionButtonRow/index.js
@@ -1,0 +1,4 @@
+import SectionButtonRow, { SpacedSectionButtonRow, BottomAlignedTypography } from './SectionButtonRow';
+
+export { SpacedSectionButtonRow, BottomAlignedTypography };
+export default SectionButtonRow;
